💄(grist) add Crisp to Grist landing page
Previously only the La Suite landing page benefited from Crisp.
For the Grist Landing page, Crisp would help users to get in touch with
the team for their first steps with Grist.
